Refactor: Standardize API to v2 structure based on analysis report
This commit implements a significant portion of the API refactoring plan outlined in the API analysis report. The primary goal was to move towards a consistent v2 API structure, deprecate and remove old/redundant endpoints, and lay the groundwork for future API development.

Key changes include:

1.  **OAuth2.1 Core Interface Standardization:**
    *   I migrated `/api/oauth/authorize`, `/api/oauth/token`, and `/api/oauth/consent` to `/api/v2/oauth/`.
    *   I migrated `/api/.well-known/jwks.json` and `/api/.well-known/openid-configuration` to `/api/v2/.well-known/`.
    *   I updated `/api/.well-known/oauth-authorization-server` to reference the new v2 endpoints.
    *   I implemented 308 permanent redirects for the old OAuth paths to maintain backward compatibility.

2.  **Permission Check Interface Consolidation:**
    *   I designated `/api/v2/users/[userId]/permissions/verify` and `batch-verify` as the standard v2 permission endpoints.
    *   I ensured `/api/permissions/check` acts as a v1 compatibility layer, calling the underlying permission service.
    *   I removed deprecated/redundant routes: `/api/auth/check-batch`, `/api/v1/auth/check`, `/api/v1/auth/check-batch`.

3.  **Basic CRUD Interface Implementation (User Management & Structure):**
    *   I created placeholder route handlers for User management under `/api/v2/users/` (list, create) and `/api/v2/users/[userId]/` (get, update, delete).
    *   I established placeholder directory structures and route files for:
        *   Role Management (`/api/v2/roles/`, `/api/v2/roles/[roleId]/`, etc.)
        *   Permission Management (`/api/v2/permissions/`, `/api/v2/permissions/[permissionId]/`)
        *   Client Management (`/api/v2/clients/`, `/api/v2/clients/[clientId]/`)

4.  **Audit and Monitoring Interface Refinement:**
    *   I migrated `/api/v1/admin/audit-logs` to `/api/v2/audit-logs/`.
    *   I migrated `/api/v1/system/metrics` to `/api/v2/system/metrics/`.
    *   I migrated `/api/health` (unversioned) to `/api/v2/health/`.
    *   I created placeholder structures for `/api/v2/audit-logs/[logId]`, `/api/v2/system/status`, and `/api/v2/system/performance`.

This refactoring provides a cleaner, more consistent API structure aligned with v2 standards and RESTful principles, addressing many of the concerns from the API analysis report. Placeholder files include TODO comments for future implementation of the actual business logic.
