+++ app/api/v2/audit-logs/[logId]/route.ts
@@ -1,0 +1,4 @@
+// app/api/v2/audit-logs/[logId]/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+interface RouteContext { params: { logId: string }; }
+export async function GET(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `GET /api/v2/audit-logs/${context.params.logId} - Get Audit Log Details (Placeholder)` }); }
