--- app/api/v2/audit/.gitkeep
@@ -1,1 +1,0 @@
-# This file is to ensure the directory is tracked by Git.
