+++ app/api/v2/clients/[clientId]/route.ts
@@ -1,0 +1,6 @@
+// app/api/v2/clients/[clientId]/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+interface RouteContext { params: { clientId: string }; }
+export async function GET(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `GET /api/v2/clients/${context.params.clientId} (Placeholder)` }); }
+export async function PUT(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `PUT /api/v2/clients/${context.params.clientId} (Placeholder)` }); }
+export async function DELETE(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `DELETE /api/v2/clients/${context.params.clientId} (Placeholder)` }); }
