+++ app/api/v2/clients/route.ts
@@ -1,0 +1,4 @@
+// app/api/v2/clients/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+export async function GET(request: NextRequest) { return NextResponse.json({ message: 'GET /api/v2/clients - List Clients (Placeholder)' }); }
+export async function POST(request: NextRequest) { return NextResponse.json({ message: 'POST /api/v2/clients - Create Client (Placeholder)' }, { status: 201 }); }
