+++ app/api/v2/health/route.ts
