--- app/api/v2/permissions/.gitkeep
@@ -1,1 +1,0 @@
-# This file is to ensure the directory is tracked by Git.
