+++ app/api/v2/permissions/[permissionId]/route.ts
@@ -1,0 +1,6 @@
+// app/api/v2/permissions/[permissionId]/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+interface RouteContext { params: { permissionId: string }; }
+export async function GET(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `GET /api/v2/permissions/${context.params.permissionId} (Placeholder)` }); }
+export async function PUT(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `PUT /api/v2/permissions/${context.params.permissionId} (Placeholder)` }); }
+export async function DELETE(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `DELETE /api/v2/permissions/${context.params.permissionId} (Placeholder)` }); }
