+++ app/api/v2/permissions/route.ts
@@ -1,0 +1,4 @@
+// app/api/v2/permissions/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+export async function GET(request: NextRequest) { return NextResponse.json({ message: 'GET /api/v2/permissions - List Permissions (Placeholder)' }); }
+export async function POST(request: NextRequest) { return NextResponse.json({ message: 'POST /api/v2/permissions - Create Permission (Placeholder)' }, { status: 201 }); }
