+++ app/api/v2/roles/[roleId]/permissions/route.ts
@@ -1,0 +1,5 @@
+// app/api/v2/roles/[roleId]/permissions/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+interface RouteContext { params: { roleId: string }; }
+export async function GET(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `GET /api/v2/roles/${context.params.roleId}/permissions - List Role Permissions (Placeholder)` }); }
+export async function POST(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `POST /api/v2/roles/${context.params.roleId}/permissions - Assign Permission to Role (Placeholder)` }); }
