+++ app/api/v2/roles/[roleId]/route.ts
@@ -1,0 +1,6 @@
+// app/api/v2/roles/[roleId]/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+interface RouteContext { params: { roleId: string }; }
+export async function GET(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `GET /api/v2/roles/${context.params.roleId} (Placeholder)` }); }
+export async function PUT(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `PUT /api/v2/roles/${context.params.roleId} (Placeholder)` }); }
+export async function DELETE(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `DELETE /api/v2/roles/${context.params.roleId} (Placeholder)` }); }
