+++ app/api/v2/system/metrics/route.ts
