+++ app/api/v2/system/performance/route.ts
@@ -1,0 +1,3 @@
+// app/api/v2/system/performance/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+export async function GET(request: NextRequest) { return NextResponse.json({ message: 'GET /api/v2/system/performance - Get System Performance (Placeholder)' }); }
