+++ app/api/v2/users/[userId]/roles/route.ts
@@ -1,0 +1,5 @@
+// app/api/v2/users/[userId]/roles/route.ts (Placeholder)
+import { NextRequest, NextResponse } from 'next/server';
+interface RouteContext { params: { userId: string }; }
+export async function GET(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `GET /api/v2/users/${context.params.userId}/roles - List User Roles (Placeholder)` }); }
+export async function POST(request: NextRequest, context: RouteContext) { return NextResponse.json({ message: `POST /api/v2/users/${context.params.userId}/roles - Assign Role to User (Placeholder)` }); }
