+++ app/api/v2/users/[userId]/route.ts
@@ -1,0 +1,37 @@
+// app/api/v2/users/[userId]/route.ts
+import { NextRequest, NextResponse } from 'next/server';
+import { withErrorHandler } from '@/lib/api/errorHandler';
+import { withAuth } from '@/lib/auth/middleware';
+
+interface RouteContext {
+  params: { userId: string };
+}
+
+// Placeholder for GET /api/v2/users/{userId} (Get User Details)
+async function getUserDetails(request: NextRequest, context: RouteContext) {
+  const { userId } = context.params;
+  // TODO: Implement actual logic to get user details
+  console.log(`API HIT: GET /api/v2/users/${userId} - Get User Details (Placeholder)`);
+  return NextResponse.json({ message: `GET /api/v2/users/${userId} - Get User Details (Placeholder)` });
+}
+
+// Placeholder for PUT /api/v2/users/{userId} (Update User)
+async function updateUser(request: NextRequest, context: RouteContext) {
+  const { userId } = context.params;
+  // const body = await request.json();
+  // TODO: Implement actual logic to update user
+  console.log(`API HIT: PUT /api/v2/users/${userId} - Update User (Placeholder)`);
+  return NextResponse.json({ message: `PUT /api/v2/users/${userId} - Update User (Placeholder)` });
+}
+
+// Placeholder for DELETE /api/v2/users/{userId} (Delete User)
+async function deleteUser(request: NextRequest, context: RouteContext) {
+  const { userId } = context.params;
+  // TODO: Implement actual logic to delete user
+  console.log(`API HIT: DELETE /api/v2/users/${userId} - Delete User (Placeholder)`);
+  return NextResponse.json({ message: `DELETE /api/v2/users/${userId} - Delete User (Placeholder)` });
+}
+
+export const GET = withErrorHandler(withAuth(getUserDetails, { requiredPermissions: ['users:read'] }));
+export const PUT = withErrorHandler(withAuth(updateUser, { requiredPermissions: ['users:update'] }));
+export const DELETE = withErrorHandler(withAuth(deleteUser, { requiredPermissions: ['users:delete'] }));
