+++ app/api/v2/users/route.ts
@@ -1,0 +1,22 @@
+// app/api/v2/users/route.ts
+import { NextRequest, NextResponse } from 'next/server';
+import { withErrorHandler } from '@/lib/api/errorHandler';
+import { withAuth } from '@/lib/auth/middleware';
+
+// Placeholder for GET /api/v2/users (List Users)
+async function listUsers(request: NextRequest) {
+  // TODO: Implement actual logic to list users
+  console.log('API HIT: GET /api/v2/users - List Users (Placeholder)');
+  return NextResponse.json({ message: 'GET /api/v2/users - List Users (Placeholder)' });
+}
+
+// Placeholder for POST /api/v2/users (Create User)
+async function createUser(request: NextRequest) {
+  // TODO: Implement actual logic to create a user
+  // const body = await request.json();
+  console.log('API HIT: POST /api/v2/users - Create User (Placeholder)');
+  return NextResponse.json({ message: 'POST /api/v2/users - Create User (Placeholder)' }, { status: 201 });
+}
+
+export const GET = withErrorHandler(withAuth(listUsers, { requiredPermissions: ['users:list'] }));
+export const POST = withErrorHandler(withAuth(createUser, { requiredPermissions: ['users:create'] }));
